Added Player Controller
After struggling with the player controller for days, I incorporated a slimmed down version of the BasicFPCC controller I was able to find online. It is sourced and cited within the script. Additionally added box collider floors to all rooms except the stairs. Stairs use a mesh collider.

I also added a .gitignore file so we can actually use github desktop with this repo. :)